requirements.txt
blessings==1.7
bpython==0.19
certifi==2020.4.5.2
chardet==3.0.4
curtsies==0.3.1
dj-database-url==0.5.0
Django==2.1.5
django-debug-toolbar==1.11
django-heroku==0.3.1
django-taggit==0.23.0
greenlet==0.4.16
idna==2.9
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
py-trello==0.17.1
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sentry-sdk==0.7.6
six==1.15.0
sqlparse==0.3.1
unicode==2.7
Unidecode==1.0.23
urllib3==1.25.9
wcwidth==0.2.4
whitenoise==5.1.0